refactor(generation): split wallPatternBuilder into focused modules
Decompose the 701-LOC wallPatternBuilder.ts into a slim orchestrator
plus three focused sibling modules. No behavior change.

- wallPatternTypes.ts (58 LOC): cache name constants + CutoutClipParams /
  HandleClipParams / RampZoneClipParams interfaces. Owned separately so
  the compound and clip modules don't form a circular dependency.
- wallPatternCompound.ts (114 LOC): buildWallPatternCompound (the
  expensive hex-prism transformCopy + compound), getCachedBaseCompound,
  buildClippedWallPattern.
- wallPatternClips.ts (236 LOC): applyWallPatternClips — the cutout,
  handle, and ramp-zone border clipping passes.
- wallPatternBuilder.ts (342 LOC): buildWallPatterns orchestrator only.
